fix(lfx): percent-encode "/" in database connection passwords
transform_connection_string encoded the password with quote(), whose
default safe="/" leaves slashes raw while every other reserved character
is encoded. Use safe="" so any valid password is fully encoded, and
collapse the duplicate langflow helper into a re-export of the lfx one.
