+++ src/backend/base/langflow/utils/connection_string_parser.py
@@ -1,8 +1,5 @@
-from urllib.parse import quote
+"""Connection string parser utilities for langflow - imports from lfx."""
 
+from lfx.utils.connection_string_parser import transform_connection_string
 
-def transform_connection_string(connection_string) -> str:
-    auth_part, db_url_name = connection_string.rsplit("@", 1)
-    protocol_user, password_string = auth_part.rsplit(":", 1)
-    encoded_password = quote(password_string)
-    return f"{protocol_user}:{encoded_password}@{db_url_name}"
+__all__ = ["transform_connection_string"]
